Send health.

+++ core/src/me/shreyasr/elemental/field/Entities/Effects/Defense.java
@@ -1,14 +1,11 @@
 package me.shreyasr.elemental.field.entities.Effects;
 
-/**
- * Created by Benjamin on 5/24/2015.
- */
 public class Defense extends Effect{
     public Defense(double powerLevel, double health){
         super(powerLevel,health);
     }
     @Override
-    public double getDef(double org){
-        return org*powerLevel;
+    public int getDef(int org){
+        return (int) (org*powerLevel);
     }
 }
